Refactor code rename dynamicWrapper to lazy
