--- src/.eslintrc
@@ -1,1 +1,0 @@
-{}
