+++ src/components/Routes/index.tsx
@@ -4,12 +4,9 @@ import { Switch } from 'react-router-dom';
 import RouteGuard from '@/components/RouteGuard';
 import RouteRedirect from '@/components/RouteRedirect';
 
-function Routes(routes, extraProps = {
-  guardData: Object
-}, switchProps = {}) {
+function Routes(routes, extraProps = {guardData: Object}, switchProps = {}) {
   if (routes) {
     return (
-      // eslint-disable-next-line react/jsx-props-no-spreading
       <Switch {...switchProps}>
         {routes.map((route, i) => {
           if (route.redirect) {
@@ -43,7 +40,6 @@ function Routes(routes, extraProps = {
                 );
                 if (route.component) {
                   return (
-                    // eslint-disable-next-line react/jsx-props-no-spreading
                     <route.component {...props} route={route}>
                       {childRoutes}
                     </route.component>
