+++ src/index.tsx
@@ -1,4 +1,4 @@
-export { default as dynamicWrapper } from '@/services/dynamicWrapper';
-export { default as history } from '@/services/browserHistory';
+export { default as lazy } from '@/utils/lazy';
+export { default as history } from '@/utils/browserHistory';
 export { default as RouterGuard } from '@/RouterGuard';
 export { Link, Route, Redirect, Router, BrowserRouter, Switch, withRouter, NavLink } from 'react-router-dom';
