--- src/services/index.ts
@@ -1,3 +1,0 @@
-export { default as dynamicWrapper } from './dynamicWrapper';
-export { default as browserHistory } from './browserHistory';
-export { default as routes } from './routes';
